Refactor server god objects into focused modules
Extract startServer message decoding/dispatch helpers and split SessionBackup utilities into dedicated modules while keeping behavior unchanged.
